vector: fix the data type of 'capacity' parameter of 'vector_alloc' function
Make it match with the data type of the data member variable of 'vector_t' user type.

+++ c/libs/vector/vector.c
@@ -53,7 +53,7 @@ static int _vector_insert(vector_t *self, int index, void *element)
         return 0;
 }
 
-vector_t *vector_alloc(int capacity)
+vector_t *vector_alloc(size_t capacity)
 {
         vector_t *self;
         size_t element_size;
